Más arreglos.

+++ eventos/en_reacciones.py
@@ -4,10 +4,7 @@ import psycopg2
 from variables import tabla_destacados
 
 async def pone_destacados(client, reaction, user):
-	chanel = reaction.message.channel
-	if user == reaction.message.author or user.bot:
-		await client.send_message(channel, "Ni tú ni los bots puedes destacar tus propios mensajes, "+
-															user.display_name+".")
+	channel = reaction.message.channel
 	else:
 		BD_URL = os.getenv("DATABASE_URL")
 		base_de_datos = psycopg2.connect(BD_URL, sslmode='require')
@@ -15,7 +12,10 @@ async def pone_destacados(client, reaction, user):
 		bd.execute(tabla_destacados)
 		bd.execute("SELECT id_canal, emoji, minimo, ids_destacados, ids_destaque FROM destacados")
 		id_canal, emoji, minimo, ids_destacados, ids_destaque = bd.fetchone()[0]
-		if id_canal == None:
+		if user == reaction.message.author or user.bot:
+			await client.send_message(channel, "Ni tú ni los bots puedes destacar tus propios mensajes, "+
+															user.display_name+".")
+		elif id_canal == None:
 			await client.send_message(channel, "Aún no han seleccionado ningún canal para mensajes destacados.")
 		elif emoji == None:
 			await client.send_message(channel, "No se ha seleccionado ningún emoji para mensajes destacados.")
@@ -48,18 +48,19 @@ async def quita_destacados(client, reaction, user):
 	base_de_datos = psycopg2.connect(BD_URL, sslmode='require')
 	bd = base_de_datos.cursor()
 	bd.execute(tabla_destacados)
-	bd.execute("SELECT canal, ids_destacados, ids_destaque FROM destacados")
-	canal, ids_destacados, ids_destaque = bd.execute.fetchone()[0]
+	bd.execute("SELECT canal, emoji, ids_destacados, ids_destaque FROM destacados")
+	canal, emoji, ids_destacados, ids_destaque = bd.execute.fetchone()[0]
 	ids_destacados = ids_destacados.split(",")
 	ids_destaque = ids_destaque.split(",")
-	if reaction.message.id in ids_destacados and reaction.count == 0:
-		i = ids_destacados.index(reaction.message.id)
-		mensaje = await client.get_message(canal, ids_destaque[i])
-		ids_destacados_new = ids_destacados.remove(reaction.message.id)
-		del ids_destaque[i]
-		await client.delete_message(mensaje)
-		bd.execute("UPDATE destacados SET ids_destacados = %s, ids_destaque = %s WHERE ids_destacados = %s", (ids_destacados_new,
-					ids_destaque, ids_destacados))
-		base_de_datos.commit()
+	if reaction.emoji == emoji:
+		if reaction.message.id in ids_destacados and reaction.count == 0:
+			i = ids_destacados.index(reaction.message.id)
+			mensaje = await client.get_message(canal, ids_destaque[i])
+			ids_destacados_new = ids_destacados.remove(reaction.message.id)
+			del ids_destaque[i]
+			await client.delete_message(mensaje)
+			bd.execute("UPDATE destacados SET ids_destacados = %s, ids_destaque = %s WHERE ids_destacados = %s", (ids_destacados_new,
+						ids_destaque, ids_destacados))
+			base_de_datos.commit()
 	bd.close()
 	base_de_datos.close()
